feat(lexer): added basic lexer code
This adds basic lexer code to Escaya, and handle diagnostics correctly on lexer level

Note! String literals are parsed as directives if first on the line e.g. "foo";

So any invalid string literals are parsed as whole instead of breaking up a invalid unicode escape and parse e.g '{' as a block statement..

This helps us to recover fast in the parser itself.

Invalid identifier escapes just output simple diagnostics and return a negative value . '-1' to avoid multiple diagnostics for the same reason.

In 'normal mode' this code will never be reached