docs(ast): respecced for statement declarations
The specs now conforms to the ECMA specs, and a 'ForBinding' have been added to reprecent the variable decl  for "in/of" bindings

ForDeclaration node removed.

There are mayor changes in the parser code, but this changes will be pushed when they are ready.

Latest changes can be tested in the online demo or through the NPM package