refactor(ast): simplified pattern with no init
If followed the specs the `BindingElement` with an optional initilizer would require the `right` property to be set to `null`.

Now it's equal to how `SingleNameBinding` are handled, and returned *as is* without wrapping in `BindingElement` if no initializer so the `null` property have been removed.

This change also outputs an AST with fewer bytes.